feat: Add `MenuButton`
