+++ packages/reakit/src/Menu/__tests__/MenuDisclosure-test.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 import { render } from "reakit-test-utils";
-import { MenuDisclosure } from "../MenuDisclosure";
+import { MenuButton } from "../MenuButton";
 
-const props: Parameters<typeof MenuDisclosure>[0] = {
+const props: Parameters<typeof MenuButton>[0] = {
   baseId: "base",
   toggle: jest.fn(),
   placement: "bottom",
@@ -13,7 +13,7 @@ const props: Parameters<typeof MenuDisclosure>[0] = {
 
 test("render", () => {
   const { baseElement } = render(
-    <MenuDisclosure {...props}>disclosure</MenuDisclosure>
+    <MenuButton {...props}>disclosure</MenuButton>
   );
   expect(baseElement).toMatchInlineSnapshot(`
     <body>
