+++ packages/reakit/src/Menu/index.ts
@@ -2,6 +2,7 @@ export * from "./Menu";
 export * from "./MenuArrow";
 export * from "./MenuBar";
 export * from "./MenuBarState";
+export * from "./MenuButton";
 export * from "./MenuDisclosure";
 export * from "./MenuGroup";
 export * from "./MenuItem";
